fix: move form context inside `EditRoleGroupModal`
