+++ src/components/[guild]/AccessHub/components/RoleGroupCards/components/RoleGroupCardMenu.tsx
@@ -1,9 +1,6 @@
 import { Box, MenuItem, useColorModeValue, useDisclosure } from "@chakra-ui/react"
-import { RoleGroupFormType } from "components/[guild]/CreateRoleGroupModal/components/RoleGroupForm"
-import useGroup from "components/[guild]/hooks/useGroup"
 import PlatformCardMenu from "components/[guild]/RolePlatforms/components/PlatformCard/components/PlatformCardMenu"
 import { PencilSimple, TrashSimple } from "phosphor-react"
-import { FormProvider, useForm } from "react-hook-form"
 import useDeleteRoleGroup from "../hooks/useDeleteRoleGroup"
 import EditRoleGroupModal from "./EditRoleGroupModal"
 
@@ -12,23 +9,11 @@ type Props = {
 }
 
 const RoleGroupCardMenu = ({ groupId }: Props) => {
-  const group = useGroup(groupId)
-  const { name, imageUrl, description } = group ?? {}
-
   const { isOpen, onOpen, onClose } = useDisclosure()
 
   const removeMenuItemColor = useColorModeValue("red.600", "red.300")
   const { onSubmit: onDeleteRoleGroup, isLoading: isDeleteRoleGroupLoading } =
     useDeleteRoleGroup(groupId)
-
-  const methods = useForm<RoleGroupFormType>({
-    mode: "all",
-    defaultValues: {
-      name,
-      imageUrl: imageUrl ?? "",
-      description: description ?? "",
-    },
-  })
 
   return (
     <Box position="absolute" top={2} right={2}>
@@ -46,14 +31,12 @@ const RoleGroupCardMenu = ({ groupId }: Props) => {
         </MenuItem>
       </PlatformCardMenu>
 
-      <FormProvider {...methods}>
-        <EditRoleGroupModal
-          isOpen={isOpen}
-          onClose={onClose}
-          groupId={groupId}
-          onSuccess={onClose}
-        />
-      </FormProvider>
+      <EditRoleGroupModal
+        isOpen={isOpen}
+        onClose={onClose}
+        groupId={groupId}
+        onSuccess={onClose}
+      />
     </Box>
   )
 }
