+++ src/components/[guild]/[group]/EditRoleGroupButton.tsx
@@ -1,9 +1,7 @@
 import { IconButton, useDisclosure } from "@chakra-ui/react"
 import { useRouter } from "next/router"
 import { GearSix } from "phosphor-react"
-import { FormProvider, useForm } from "react-hook-form"
 import EditRoleGroupModal from "../AccessHub/components/RoleGroupCards/components/EditRoleGroupModal"
-import { RoleGroupFormType } from "../CreateRoleGroupModal/components/RoleGroupForm"
 import useGroup from "../hooks/useGroup"
 import useGuild from "../hooks/useGuild"
 import { useThemeContext } from "../ThemeContext"
@@ -13,18 +11,8 @@ const EditRoleGroupButton = () => {
 
   const { urlName } = useGuild()
   const group = useGroup()
-  const { name, imageUrl, description } = group ?? {}
 
   const { textColor, buttonColorScheme } = useThemeContext()
-
-  const methods = useForm<RoleGroupFormType>({
-    mode: "all",
-    defaultValues: {
-      name,
-      imageUrl: imageUrl ?? "",
-      description: description ?? "",
-    },
-  })
 
   const { isOpen, onOpen, onClose } = useDisclosure()
 
@@ -40,16 +28,14 @@ const EditRoleGroupButton = () => {
         onClick={onOpen}
       />
 
-      <FormProvider {...methods}>
-        <EditRoleGroupModal
-          isOpen={isOpen}
-          onClose={onClose}
-          groupId={group.id}
-          onSuccess={(response) => {
-            router?.push(`/${urlName}/${response.urlName}`)
-          }}
-        />
-      </FormProvider>
+      <EditRoleGroupModal
+        isOpen={isOpen}
+        onClose={onClose}
+        groupId={group.id}
+        onSuccess={(response) => {
+          router?.push(`/${urlName}/${response.urlName}`)
+        }}
+      />
     </>
   )
 }
